requirements.txt
streamlit>=1.35.0
streamlit-cropper>=0.2.1
ultralytics>=8.2.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
Pillow>=10.0.0
pinecone-client>=3.2.0
python-dotenv>=1.0.0
numpy>=1.26.0
opencv-python-headless>=4.9.0